RAG/requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1

# Database & Storage
asyncpg>=0.29.0
supabase>=2.3.0
upstash-vector>=0.4.0

# AI & ML
google-generativeai>=0.4.0
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.30.0

# Document Processing
pypdf>=4.0.1
langchain-text-splitters>=0.0.1
